Eliminate instantiations of apache::thrift::detail::st::ComparisonOperators
Summary:
Build profiling showed that instantiations of `apache::thrift::detail::st::ComparisonOperators` was a significant source of slow-down at compile time. `ComparisonOperators` is a helper that gives Thrift-generated classes four relational operators, each of which is a one-liner. Instead of using a template for this, just give each Thrift class those four operators directly.

In addition, define an `apache::thrift::detail::st::IsThriftClass` type trait, for detecting whether a type is a Thrift-generated class. This is for use in places where code was testing for derivation from `ComparisonOperators` as a hackish way to detect the same trait.

Reviewed By: yfeldblum, iahs

Differential Revision: D22055804

fbshipit-source-id: 18d00c5fa991b2e065462f1f87c725082df0abbe