Expose isset for consumption in python
Summary:
There currently is not an easy way to know if a field was set or just using the default value.

Introducing Struct.isset(obj).field -> bool

As to not pollute the Struct instances with _ prefixed methods, and we don't want to train people to expect that they can get away with using _ prefixed methods on thrift objects.  That is all reserved as implementation detail.

I used a cdef on each struct so its only visible in Cython, named __isset so its impossible to conflict with without first conflicting in cpp2.

I decided to attach the isset inspection function to the MetaClass of Struct so we don't further pollute the thrift.py3 namespace with yet another inspection function.  And it won't show up on instances of Struct this way.

Created a GeneratedError class that parents all generated Thrift Exceptions.

Struct.isset() works with GeneratedErrors. In Python, a GeneratedError is not a subclass of Struct. In Thrift they are so we can get away with it here.   GeneratedErrors would be subclasses of Struct if that were possible, but that's not allowed by Cython (no multiple inheritance), and they HAVE to be subclasses of Error.

Reviewed By: fried

Differential Revision: D18610445

fbshipit-source-id: c42b07f679cdee83501e951c8d9eb8f3eb746618