Add variants function to enum generated from thrift for use with structopt
Summary: Previously you would have to wrap the type we generate from the thrift yourself if you wanted this and then impl FromStr and variants yourself or use arg_enum. With this change you don't have to, you can just use the thrift generated type as if it were an enum wrapped with `arg_enum`.

Reviewed By: dtolnay

Differential Revision: D22722303

fbshipit-source-id: 15e8a0a00a213372e56a8ad5384a3e025bb83936