Support for allocator aware C++ structs
Summary:
Create two new annotations to allow Thrift-defined C++ structures to be "allocator aware":
1. `cpp.allocator` (per struct)
2. `cpp.use_allocator` (per field)

`cpp.allocator` is used on a struct definition, indicating that we wish to declare an allocator-aware struct.  It takes the allocator type as an argument.  When this annotation is present, we generate the customary `allocator_type` typedef, as well as the 3 "allocator extended" constructors:
- `Foo(allocator_type)`
- `Foo(const Foo&, allocator_type)`
- `Foo(Foo&&, allocator_type)`

`cpp.use_allocator` is applied to individual fields of an allocator-aware Thrift struct to indicate that we want to pass our allocator on to this field.  This makes it possible to gradually incorporate allocator awareness into existing code.

See the new unit test for a demonstration.

#forcetdhashing

Reviewed By: davidtgoldblatt

Differential Revision: D22797309

fbshipit-source-id: 4b55624279ff0fc396a2e095d73411657223de91