Add support for `peekList/Set/Map` in generated deserializer
Summary:
The generated deserialized code currently expect the container to be sized in
advance, i.e. we should know before deserializing the values, the size of the
container. This is only true for TBinary and TCompact serialization protocol
but not for other serialization format like SimpleJSON.

This diff add a new behavior in the deserialization code, which now check for
the size of the container, if it's < 0, it uses the `peekList/Set/Map` to know
if it reaches the end of the container.

Reviewed By: yfeldblum

Differential Revision: D22951463

fbshipit-source-id: f488cd074de47cc05f845abd2966513ea21de9a6