handle build() method for big struct correctly
Summary:
D22994891 exposed an issue with current thrift compiler for java. When a struct is big, we double apply build method in generated code, leading to multiple return statements in generated code.

In this diff, we make sure if a struct is indeed big, only one code generation path is used.

Reviewed By: prakashgayasen

Differential Revision: D23012263

fbshipit-source-id: 0d6e2fdcc6f31fd75dd97cebec57e3746e6ec7e8