remove default implementation of ThriftMetadata generator
Summary:
# Current logic

Currently thrift metatdata generator is template class with default implementation that does nothing:

https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/thrift/lib/cpp2/gen/module_metadata_h.h?commit=3f225926a10a0b734ab01333691765317e0b863d&lines=34-38

We use codegen to generate template specialization for each thrift structures

https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/thrift/compiler/generate/templates/cpp2/module_metadata.h.mustache?commit=3e19ff57ef31e151202b88126b522c60a203ea5b&lines=49-53

If `no_metadata` option is turned on, we do not generate specialization, so that the default implementation wil be used.

# This diff

The problem is that if people inadvertently included template generator without specialization version from codegen, they will get empty data, even with `no_metadata` turned off.

This diff remove default implementation of ThriftMetadata generator, so that we get compile-error in this case.

Reviewed By: vitaut

Differential Revision: D22984252

fbshipit-source-id: 54ea89a95a528784708c14a8eeb2fcc0088f38a2