C++: Fix bug in TCompactProtocol while reading the message type
Summary:
In the `TCompactProtocol`, we encode the version number and the message type on
one single byte. 5 bits for the version number and 3 bits for the message type
like this `VVVVVTTT`.
`TCompactProtocol` uses the `0000_0011` binary mask (`0x03`) to extract the
message type information from the byte where it should have been `0000_0111`
(`0x07`). The bug is only visible with message type bigger than 3, and only
`T_ONEWAY` (`= 0x04`) is in this case.

However, by chance this bug doesn't manifest itself in the code (so far), since
AFAIK we only check for `T_EXCEPTION` and `T_REPLY` in
`thrift/lib/cpp2/GeneratedCodeHelper.h`, not `T_ONEWAY`).

Reviewed By: yfeldblum

Differential Revision: D23054525

fbshipit-source-id: 836e7c9d182b65837e9fc221fb3e35e09a007796