fix queued_requests counter
Summary:
Thrift exposes `queued_requests` counters. There are two varieties of those: one is a snapshot  (current value), the other one is an average of snapshots taken throughout some time period (specifically, when any request is read from the wire).

the problem is that we were showing not number of pending requests, but number of tasks in TM's queues. This does not correctly reflect what this counter is supposed to be about.

Fix is to only count the elements in the "upstream"  queues.

Reviewed By: yfeldblum

Differential Revision: D23013115

fbshipit-source-id: 7e5438ab4ee775df174f7c70f3e3268895add9cb