Remove implicit operator= binding from any_ref and add assign_value function
Summary:
Adds an explicit `assign_value` and `try_assign_value` function for assign
It takes a non-trivial amount of work to support assigning to an std::any in addition to a normal instance variable. As this is one of the goals of any_ref, it makes much more sense to build this in, then have sporadic support for the feature.

Additionally, the previous version was prone to bugs where a reference was captured instead of a value being assigned. Now any_ref works almost exactly like a normal reference for assignment. The only remaining difference is that the any_ref copy constructor can be used to explicitly request the reference change.

I've also:
- made `has_value` and `bool()` match std::any semantics.
- added `has_reference`, which returns true even when referring to an empty std::any.

Reviewed By: yfeldblum

Differential Revision: D23086830

fbshipit-source-id: 20019fc6919c4bf5a662e752a16a824d4f3271ef