Java: Fix bug in TCompactProtocol reading of message type
Summary:
TCompactProtocol encodes the version and the type on a single byte, 3 bits for
the type and 5 bits for the version (i.e. `TTTVVVVV`).
The VERSION_MASK is `00011111` and the type mask should be (after 5 bits
shifting) `00000111` (7) not `00000011` (3).

The bug was not very obvious because, the types are defined like this:
```
CALL = 1;
REPLY = 2;
EXCEPTION = 3;
ONEWAY = 4;
```
And because, `ONEWAY` is rarely used, it wasn't obvious.

Reviewed By: yfeldblum, prakashgayasen

Differential Revision: D23059405

fbshipit-source-id: d9a2311d584c334528f90a9366f228882fc5330a