Java: Fix minimum size of field in TCompactProtocol
Summary:
The `typeMinimumSize` method in `TCompactProtocol` is used to find the minimum
amount of data needed to represent a type.
There was a bug indicating that `float` needed 4 bytes and `double` 8 bytes
(like `TBinaryProtocol`), but this is incorrect.
Both `float` and `double` are converted into their Integer/Double
representation before being encoded using variable length. So the actual
minimum is 1 bytes in both cases.

Reviewed By: yfeldblum

Differential Revision: D23179565

fbshipit-source-id: b58070251044a56c5a5162d460246d59f63dff04