reject string on list<string> fields
Summary:
We used to accept `string` value on a field with `list<string>` type. Because string is a sequence type in Python, we implicitly convert the string to a list of single character strings.

But that behavior is quite error prone. This diff requires explicit conversion to avoid surprise.

Same applies to `(string|binary)` -> `(list|set)<(string|binary)>`.

Reviewed By: yfeldblum, fried

Differential Revision: D23196656

fbshipit-source-id: 78e36b07d7c33c7e482dd7964ca0d0b714962359