directly use Protocol enum from cpp2
Summary: `thrift.py3.common.Protocol` is a separate enum and not consistent with `apache::thrift::protocol::PROTOCOL_TYPES`, and we will need to use the convert function to convert the former to the latter. This diff directly cimport `PROTOCOL_TYPES` as `Protocol` so we don't need to maintain the extra enum and saves the conversion.

Reviewed By: yfeldblum

Differential Revision: D23438427

fbshipit-source-id: b70434c90af61d312f7843f554c72239c8de525a