Change channel interaction API to remove synchronization requirement
Summary:
Requiring the caller to decide whether to use setNewInteraction or setExistingInteraction forces synchronization in clients accessed from multiple threads. But we already have to synchronize eventually to get onto the IO thread, so we hand off the interaction name to the channel and let it figure out what metadata to set.
This introduces a map in RCC, which is unopinionated and so has nowhere to store the name, but that map is empty in the fast path where all interactions have been used at least once

Reviewed By: yfeldblum

Differential Revision: D23012884

fbshipit-source-id: 48d7d522bacb942016f25b5ae4309ca8110a0390