Fixing HeaderServerChannel to handle 1-way requests with in-order responses
Summary: Usecase where this is needed: when in-order responses are needed (ie, out-of-order response flag is not set), HeaderServerChannel (https://fburl.com/diffusion/tgibgftk) will stop sending any responses once it receives a 1-way request because it waits for all responses to be sent in order.  It uses an internal sequence id to keep track   of all incoming requests and will "park" any responses until the one with the right sequence id is ready to be sent.  It does this for 1-way requests as well, and since 1-way requests have no responses, it will never send any responses back after receiving a 1-way request.

Reviewed By: andriigrynenko

Differential Revision: D23489244

fbshipit-source-id: 5114b0fae27ce62800db8e00fdd4bd1902e19f8e