to_py3_struct should respect py3.name annotations
Summary:
We have a union that has a py3.name annotation on a field,
which results in a crash trying to convert a py version of the struct
to py3.

This adds handling for this case by preferring the name in py3.name if
present when converting a struct or a union.

Reviewed By: yfeldblum

Differential Revision: D23643026

fbshipit-source-id: 712e286c2ec82cb40642c2e46f2883d230e76782