Add generators for 'interesting' and 'key' values.
Summary:
Key values are interesting values that are unique and can be used in a set or as a map key.

Note that this is complementary to thrift/lib/cpp2/reflection/populator.h, which supports structs and generates random values (while this does not support structs and generates hand selected 'interesting' values). Ultimately they should be merged and give you the choice of random, interesting, and key values.

Reviewed By: iahs

Differential Revision: D23301513

fbshipit-source-id: 02c55f32863ec8cf15caae09f66049a7df3eaa5c