Enable client to pass TOS to reflect in setup frame metadata
Summary:
This diff allows individual client connections to send a hint in the rocket setup frame that the server should apply a particular DSCP value to its end of the connection. (In practice, this will be the same DSCP value the client set on its end.)

If the server has enabled global TOS reflection, the new logic need not be applied.

Reviewed By: yfeldblum

Differential Revision: D23488594

fbshipit-source-id: a241b00a98e266861305a3cb9f48477d1d35fd85