thrift/rust: allow mock closures to return Err variants
Summary:
We have a little mocking framework in our Thrift codegen, which lets us create
a mock client for use in tests. This is convenient, but right now it cannot
return an Err variant from a closure, so that means you can only return errors
through `.throw()`, but this only works for errors that are `Clone`.

Let's relax this a bit by adding a `mock_result` helper that lets us pass any
closure to stub the function (we already had `mock`, but that can only return
an OK variant).

#forcetdhashing

Reviewed By: StanislavGlebik

Differential Revision: D23710875

fbshipit-source-id: 1fc5bf74f690212a6e67ea454edc26c0e04aee9a