Destroy tiles in TM thread when connection is closed
Summary:
During normal operation the generated client will send termination before closing the connection, which destroys the tile in the right thread. But in case of a crash or network partition the connection will be closed without first receiving the termination, causing the tiles to be destroyed inline on the eventbase. The tile destructor is arbitrary user code and can be responsible for cleaning up resources, so this is undesirable.
There is an interesting edgecase with TilePromise because the continuations keep the connection alive, so we have to wait for the constructor to complete before destroying them. We don't execute the continuations as an optimization, but I'm not certain whether users would expect that to be the case (espcially in case of a oneway method)

Reviewed By: yfeldblum

Differential Revision: D23659700

fbshipit-source-id: 7c38832749addec365c62096e6023577a329261b