removing __bool__ in types
Summary:
In thrift-py3 types, `__bool__` is implemented to make casting a struct object to bool type possible. The logic is:
1. if any default values are configured in any of the fields, True
2. if any optional values are set, True,
3. if no fields, True.
4. Otherwise False.

It's a good/reasonable concept but there is no (correct and intentional) usage of this feature in the code base and the logic behind this is confusing and may leads to unexpected outcomes when used without this knowledge. A typical risk use is when `foo` can be a struct object or `None`, instead of use `if foo is not None:`, users typically use `if foo:`.

It also contributes to additional compile time especially for large thrift files with a large number of structs.

Reviewed By: yfeldblum

Differential Revision: D23750646

fbshipit-source-id: 0aa738fd23dc46d82475e0062ac1cb923f952d61