Fix inconsistency in TSimpleJSONProtocol.readMessageBegin
Summary:
`TSimpleJSONProtocol.readMessageBegin` currently write the version number of
the protocol as the first element of the serialization (mimicking the behavior
of other protocols). However `writeMessageBegin` doesn't read this value,
making the protocol improper for usage a serialization protocol.
This diff fixes the problem by taking the stance that version number should be
omitted, following the spirit of JSON.

Reviewed By: avalonalex

Differential Revision: D23804082

fbshipit-source-id: 8c1d205325977a35148dbc96e138179f4f123959