Treat first 32 bits of the setup frame as a protocol key
Summary:
Currently the first 32 bits of the setup frame are treated as a major and minor version, but it's not very usable for versioning, since it doesn't allow client to pick a newer version (without reopening a connection). We also have a separate versioning mechanism inside of the RequestSetupMetadata.
This diff suggests treating these 32 bits as a unique rocket protocol key, to allow distinguishing rocket connections from other RSocket connections. For now the server will have to support both keys for compatibility.

Reviewed By: rhodo

Differential Revision: D23698321

fbshipit-source-id: 6a9eb755e5ef54c3c3c6c55ca506174c81b0ffb5