Support Rust map field having default value []
Summary:
This diff allows a field whose type is `map<k, v>` to have a default value of `[]` meaning an empty map. This is in addition to the existing support for having `{...}` mean a map, whether empty or not.

```
struct TPostingListItems {
    7: map<i16, string> payload = [],
}
```

This is just like the code 10 lines higher inside `is_set` which permits both `[...]` (with any number of values) and `{}` (as long as it's empty) to initialize a set.

Reviewed By: yfeldblum

Differential Revision: D23871870

fbshipit-source-id: e6f9f63372e20edf735bc974e1f934377a7cc629