Default the Transport/Protocol to Header
Summary:
`ThriftServerConfig` define default values for transport/protocol as
`"framed"`/`"binary"`. But 99.9% of the time, people wants `"header"` instead
since it's the default values used by other languages.

Changing the default to what most FB projects wants and removing the setting in
our user's projects allow us to take control of the parameter which is going to
be useful for the RSocket migration.

Reviewed By: yfeldblum

Differential Revision: D23510914

fbshipit-source-id: 5137d2dfb367f88a083eb1819e1691ca18891fe7