Make Thrift unions work with Go.
Summary:
Go code gen represents unions as structs and tries to access the fields of such struct even if it's optional.
The esiest way to fix it (although definitely not the prettiest) is to check for nil pointer in IsSet* helpers.
We also need change the check `CountSetFieldsUnion2()` check from `!= 1` to `> 1`, because no fields can be set if
the structure is optional.

Reviewed By: yarikk, dlespiau

Differential Revision: D23905167

fbshipit-source-id: 72f5607a9a1d970c4a88100915c60c299139bcbb