Fix LoggingEventTest to have deterministic behavior
Summary:
Right now in LoggingEventTest, event handlers are leaked with event registries, but gmock objects are required to be destructued to check expectations attached on them.
For each test fixture, we firstly set up a new mock registry storage object, so that we never reuse registries, and also we manually check the expectation on each event handler to make test deterministic.

Differential Revision: D24029769

fbshipit-source-id: 3940c68875df9a02f9f2f50efefd74bcc161d376