Change ownership model of tiles
Summary:
The current ownership model has a fun edge case where if an interaction id is reused while the first constructor is running (e.g. create on oneway method -> terminate -> create and malloc reuses the state struct's memory) the create will be registered as a duplicate id even though the contract is being followed.

The new ownership model is slightly more involved: terminated interactions are removed from the map and released, with an owning pointer in the EventTask. The task handles cleanup if it is destroyed without running (e.g. expires), and the existing decref takes over if the task is run. This removes some special casing that was required for  TaskPromises, and once ErrorTiles are replaced we'll have the nice property that the map holds only active interactions.

Reviewed By: bithree

Differential Revision: D24142339

fbshipit-source-id: 95326aedfea6108c31597575ff7d2505de1bedd5