port stopListening() over to Python interface
Summary:
We would like to perform some graceful cleanups, after SIGTERM but before server is completely killed. The first thing in the signal handler callback we would like to do is to stop new connections.

This is possible in C++ via `stopListening()`, but Python has not ported it yet. Submit this diff for comments/concerns.

Reviewed By: nanshu

Differential Revision: D24266195

fbshipit-source-id: 109ca90b48456874ef50fc5ad70d8afd22afa996