Correctly handle unknown fields during read
Summary:
Currently readField() will return `null` which is out of spec and
causes crashes. Here instead we peek the correct type and return id 0. Same as
TCompactJsonProtocol.

We copy the approach of peeking from TJSONProtocolBase. Ideally we should
rebase the protocol on that base class but that's outside of this diff.

Reviewed By: stevegury

Differential Revision: D24306515

fbshipit-source-id: 6339b2383f9b0a4ae6c8e7c87a455c4cbf19223c