Extract setup logic shared between tm and eb threading models
Summary: Currently the processing call chain has a slightly different structure for thread = tm/eb, with common setup logic stuck in the first function of each chain. That logic is currently just a checked call to validateRpcKind, but in order to support interactions using the eb threading model it will grow substantially. To reduce duplication and added complexity the two call chains are first merged in this diff, extracting a common setUpAndProcess function per method and using that in the processmap unconditionally.

Differential Revision: D24447782

fbshipit-source-id: 90ec2fb759d10e0c90ec196f0942cbac747dbe19