refactor generator to support more complex generation logics
Summary: We used to have a simple pattern for source file generation (i.e. pyx, pxd and pxi), and then we introduced wrappers (wrappers.h, wrappers.cpp and wrappers.pxd), and then types.h, and then builders and reflections. With more of these additions, it's get harder and harder to make the existing generation logic correct. Instead of using a lot of nested loops and conditionals, it's better to use a module config that explicitly list the files of different kinds (i.e. cython, cpp, skip when no services, type context etc) and then use it to make the generation logic clear.

Reviewed By: yfeldblum

Differential Revision: D23891359

fbshipit-source-id: 00c033694fd4c781a12bbcf7dc309e03763d4c38