Add cpp2 code gen for Any
Summary:
#forcetdhashing

Adds static-init time registration for a struct that has an any_type.name annotation.

An 'any' build option must be specified for this to work. This option causes:
- `*_types.cpp` to define a static init function for structs with an any type name.
- `*_types` target to depend on the AnyRegistry
- `*_sinit.cpp` to be created, whose sole purpose is to isolate the static initialization call so link_whole=True is optimal. It has no includes, and simply references the static init variable defined in `*_types.cpp`.

Ideally we could use the [[used]] attributed, but most linkers do not support it. Instead, we uses the link_whole flag on the sinit target, which is specially crafted to eliminate the down sides of using link_whole, by only including the static init variable, thus having the exact same result as using the [[used]] annotation.

The 'any' option is needed as the AnyRegistry dep cannot be added to all cpp2 targets, as that would create a circular dep for the Any struct code gen. Ideally we will default this on in the future, and explicitly disable it for 'any.thrift'. For now, you have to explicitly opt in.

If two structs define the same name, when the program is run, you get the following output:
  STDERR:terminate called after throwing an instance of 'std::runtime_error'
    what():  Could not register: facebook.com/thrift/test/MyAnyStruct

Reviewed By: iahs

Differential Revision: D24039740

fbshipit-source-id: 518542ba32c974e5073e575e4bc0d8ba7d768724