Make protocol.skip deal with container of unknown size
Summary:
`TSimpleJSON` is a special protocol which doesn't encode the same level of
information as Binary or Compact. In particular, it doesn't encode the length
of a container (e.g. `{my_list: [1, 2, 3]}`).
So while parsing the data, you can't know in advance the size of the list, we
work around this by using a `peekList` method which will check if the list
contains more elements.
The `TProtocolUtil` abstraction, used for skiping unknown fields, didn't
support that feature. This is now fixed with this diff.

Reviewed By: avalonalex

Differential Revision: D24422473

fbshipit-source-id: f23ecd524d1d6bd7b921893e298bcc6c34d76d5a