EventHandlerBase.h: make getContextStack public
Summary:
I'd like to add support for calling the ContextStack from Rust Thrift codegen,
but that doesn't get executed in a subclass of EventHandlerBase.h, though we do
have one around.

Looking at the history of this in D22193209 (https://github.com/facebook/fbthrift/commit/947cf76d723c77cbe08a793a57eab04b46fa824d), it looks like this wasn't
necessarily intended to be protected, so I hope this is OK to update it to be
public!

More context and my proposed plan in this Thrift users post:
https://fb.workplace.com/groups/thriftusers/permalink/622271158482300/

Reviewed By: yfeldblum

Differential Revision: D23628253

fbshipit-source-id: d8ce48f6ee4804c5689e1974193589a67a3c6231