support thrift includes
Summary:
when a thrift file A include another thrift file B, `A.h` will include `B.h`.
In that case, the tablebased metadata in `A.h` will reference the tablebased metadata in `B.h`.
We need to forward declare relevant metadata in `B.h` in order to compile successfully.
To help with compilation speed, we split the necessary types for forward declaration from `TableBasedSerializer` into `TableBasedForwardTypes`.

Reviewed By: vitaut

Differential Revision: D24071277

fbshipit-source-id: 3c163c5d80f825699834033e3ef42146be2baea6