Remove ErrorTile
Summary:
The interaction constructor is not meant to throw, but we need to do something if it does. We introduced ErrorTile to ensure the client would receive interaction constructor errors on all subsequent requests. But the most common usage patterns for the client, sequentially co_awaiting methods and concurrently awaiting with collect*, only surface the first exception thrown - the remaining ones won't be observed at all. This diff removes ErrorTile, which causes methods not received by the server at the time the constructor throws to receive an unknown id exception instead. We previously introduced error codes to mark these two exceptions so we can eventually rewrite the later exceptions to match the earlier ones and even retry the whole interaction because we know no methods were executed.
What we get in exchange for these error messages is the removal of a concept from the server code and the nice invariant that the contents of the tile map exactly map to the set of active interactions.

Reviewed By: bithree

Differential Revision: D23869158

fbshipit-source-id: 637d69fe0aacc819feda360657f6e8a1e9967769