Keep interaction alive until streams and sinks are completed
Summary: Terminating an interaction signals to the server that it may clean up resources once outstanding requests are completed. Currently we wait for all handler methods to return before destroying the tile, but this is too narrow a definition of completed in the case of streams. Adds a pointer to both stream implementations to defer the releaseRef call until the stream has completed. This is easy to define for the generator; for the publisher we wait until the completeOrcancel callback has been called.

Reviewed By: rhodo

Differential Revision: D24312147

fbshipit-source-id: b9e07da3471dffb92085fe02752e293e677c1a4c