Fix bug while reading a union with TJSON protocol
Summary:
A union is serialized as a struct containing only one field.
The current code use `readFieldBegin()` + `readFieldEnd()` to consume the
`TType.STOP` byte after the first field. This is ok in most protocol since
`readFieldEnd()` is a no-op in most protocol. However it causes some trouble
with `TJSON` (not to be confused with `TSimpleJSON`).

This diff remove the call to `readFieldEnd()` as it is not always a no-op.
Also, I added a comment which explain why we need to call `readFieldBegin` right after a `readFieldEnd` (which may seem like a bug at first).

Differential Revision: D24499759

fbshipit-source-id: 67f078c0a75442b3fca1c98651fc97ad00c16c6a