Fix bug while reading a union with TJSON protocol
Summary:
A union is serialized as a struct containing only one field.
The current code use readFieldBegin() + readFieldEnd() to consume the
TType.STOP byte after the first field. This is ok in most protocol since
readFieldEnd() is a no-op in most protocol. However it causes some trouble
with TJSON (not to be confused with TSimpleJSON).

This diff remove the call to readFieldEnd() as it's not always a no-op.

Differential Revision: D24499758

fbshipit-source-id: ac3ea8c36481f0d8c546e45b4f68600c9e9292aa