Use Enum.fromInteger(0) as default value for an enum
Summary:
Currently, we use `null` as the default value for an enum. It is correct when
an enum doesn't have an entry for 0. However, when a enum has a value for 0,
its default value should be this entry.

E.g.:
```
enum Color {
  GREEN = 1,
  BLUE = 2,
}

enum Taste {
  SALTY = 0,
  SWEET = 1,
}

struct Foo {
  1: Color color,
  2: Taste taste,
}
```

The default value of `Foo.color` must be `null` while the default value for
`Foo.taste` must be `SALTY`.

#forcetdhashing

Differential Revision: D24527828

fbshipit-source-id: 72d09eab7ec5c6512afa9b14436bff6ee95ed862