Move thread management outside of InMemoryConnection
Summary: Make `InMemoryConnection` not own the thread that drives the event base because due to delayed destruction the connection destructor is occasionally run on the event base thread. This was causing a "Resource deadlock avoided" exception and a crash when trying to join the current thread as well as a data race in `InMemoryConnectionTest` because nothing blocked returning from `main` and `InMemoryConnection` destructor was racing with global deinitialization.

Reviewed By: yfeldblum, glowinglake

Differential Revision: D24689055

fbshipit-source-id: d1b0a0bdacc3cb6253e89f44c671e491d28f53cd