Bugfix typo in allocator test
Summary:
Accidentally put an underscore here instead of a dot. AA-c'tors were not actually forwarding the allocator to this member, but it didn't break the test because the default-constructed allocator still throws.

Here I fix the bug, and also improve the test such that it would catch this.  (99% of the diff is renaming the allocator as "Maybe" and including a default-false "armed" member)

Before:
```
explicit UsesAllocatorChild(const allocator_type& alloc) noexcept :
  __fbthrift_alloc(alloc),
  aa_list(alloc),
  aa_set(alloc),
  aa_map(alloc),
  aa_string(),     // LOOK HERE
  not_a_container(),
  not_aa_list(),
  not_aa_set(),
  not_aa_map(),
  not_aa_string() {}
```

After:
```
 explicit UsesAllocatorChild(const allocator_type& alloc) noexcept :
   __fbthrift_alloc(alloc),
   aa_list(alloc),
   aa_set(alloc),
   aa_map(alloc),
   aa_string(alloc),   // LOOK HERE
   not_a_container(),
   not_aa_list(),
   not_aa_set(),
   not_aa_map(),
   not_aa_string() {}
```

Reviewed By: yfeldblum, therealgymmy

Differential Revision: D24758924

fbshipit-source-id: 117bfd1a881637b1ef641385801b334980965c03