Generate fromMap_DEPRECATED for all structs with mapconstruct
Summary:
With this change, `fromMap_DEPRECATED` would always be available for structs in packages, which specify the `mapconstruct` option. Without `mapconstruct` it would be just a trivial wrapper. With `frommap_construct` the constructor would switch to the positional one, as it already does.

This would be useful for a global codemod, where it can't switch to `fromShape` due to having a variable parameter rather than an inline constant expression. Here's a tricky example https://fburl.com/diffusion/qhr3gfx1, where even a manual update may be non-trivial. A global `fromMap_DEPRECATED` would allow a codemod in one sweep with a global enabling of `frommap_construct` later in one sweep as well (in fact, the option itself could be deprecated then due to always having a positonal constructor).

Reviewed By: prasad223

Differential Revision: D24807142

fbshipit-source-id: 99ac625a84249541d3aa3e0f90e31b6bd0c7c304