Use coalesce for default values in a map constructor
Summary:
This would replace
```
$this->field = idx($map, 'field', FIELD_DEFAULT);
```
with
```
$this->field = idx($map, 'field') ?? FIELD_DEFAULT;
```

The former won't use the default if the key is present but equals to null. This is inconsitent with a positional constructor and a general Thrift Hack approach of null being equivalent to not set.

This issue arises when using `fromShape` with `mapconstruct`, since it passes all fields to the constructor even if they were absent. This would change the behavior and cause tests like `AssistantStudioDeploymentStatusUtilsTest` to fail.

Reviewed By: prasad223

Differential Revision: D24808290

fbshipit-source-id: 7f572530627c851266105ba22d7cb9fc14562e7d