Add a default construction API for structs
Summary:
There are many callsites of an empty constructor, including Thrift generated code. This diff adds a `withDefaultValues` method, which is a thin wrapper over a constructor call with no arguments.

This is better than `fromShape()` for a few reasons:
* `fromShape` with nor arguments is slightly confusing ("but where is the shape?").
* `withDefaultValues` clearly describes the semantic. It might even incentivize developers to migrate away from an approach with creation of an empty object and manually filling in fields ("using default values is an extra unnecessary operation - is there a way to avoid it? (yes, there's fromShape)").

`fromShape` calls are updated to pass an explicit argument and would be updated to `withDefaultValues` once it's globally available.

Reviewed By: yfeldblum

Differential Revision: D24867333

fbshipit-source-id: 8d634ae24e9662ca9d6aba00be2e4ed550b3b2e1