Fix the optional check in mapconstruct
Summary:
If the map's key is set and the value is null, optional field may get assigned a default value as if it were unqualified.

This worked before ?? was introduced for mapconstruct, but that was fixing another inconsistency.

This fix should set a proper behavior - if an optional field is unset in map or equals to null, it stays null. If an unqualified field is set in map and equals to null, it gets assigned a default value.

Reviewed By: prasad223

Differential Revision: D24926737

fbshipit-source-id: 128a6a95ad9adf33d1996cd2115445b90ab4003a