Add a conformance test runner.
Summary:
Adds a gtest-based conformance test runner, build rule, and primary conformance test target.

TestSuites are generated programmatically by a binary. The generator must output a binary-encoded TestSuite to stdout. Test suites are always loaded entirely, regardless of what tests are actually run. The difference between running a generator vs linking in the generator appears to be ~.07 seconds (~8% for the round trip tests). In a subsequent diff, the generator binaries will also be used to create golden files of their output.

To enable conformance tests for a particular thrift implementation, implement a server that that exposed the ConformanceService on a localhost port. Then add the server target in the `servers` arg of the thrift_conformance_test rule. The server must write the localhost port number to use followed by a newline to stdout. The test runner will lazily start and connect to the server, when it is needed by a particular test case.

All the normal gtest flags work, the conformance test is just a normal gtest binary, so tests cases can be listed, filtered, etc. To run all the conformance tests use:
  buck test thrift/conformance:

It is not expected that every implementation will pass every conformance test, so text file listing non-conforming test cases can be specified. Nonconforming test cases are is not disabled, but rather the test runner will verify that it fails. The the nonconforming list should increase when we add new features and tests, then be ever shrinking as we improve conformance.

Right now NaN tests are non-conforming, not because anything is broken, but because the test runner is checking for equality not equivalence (and NaN == Nan -> false).

In this diff, I have added a generator binary for roundTrip tests, and used the cpp2 conformance test server to seed an initial set of conformance tests.

Note that buck runs each test separately, so all the test suites are loaded, and one of the servers is started per test case. If the test loading becomes a problem, we can break out large test suites into different thrift_conformance_test targets.

Here is some example output:

      ✓ ListingSuccess: thrift/conformance:round-trip-test - main (0.138)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/min.Binary (0.633)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/byte/max.SimpleJson (0.816)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/neg_denorm_min.Compact (0.817)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/byte/neg_one.SimpleJson (0.837)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/double/inf.Compact (0.838)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/string/upper.Binary (0.837)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/binary/spaces.Compact (0.860)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/double/neg_epsilon.Compact (0.864)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/double/zero.Binary (0.814)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/double/neg_one.Compact (0.888)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/neg_max_digits.Compact (0.903)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/double/neg_inf.Binary (0.904)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/string/lower.Compact (0.903)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/string/utf8.SimpleJson (0.909)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/i16/one.Binary (0.918)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/binary/upper.SimpleJson (0.918)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/neg_inf.Binary (0.836)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/byte/neg_one.Binary (0.950)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/binary/upper.Binary (0.950)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/max_digits.Compact (0.950)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/string/empty.Compact (0.971)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/binary/leading_space.Binary (0.900)
      ...
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/binary/bad_utf8.Binary (0.656)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/binary/leading_space.Compact (0.639)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/i32/min.Binary (0.702)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/binary/trailing_space.Binary (0.667)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/max_int.SimpleJson (0.641)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/min.SimpleJson (0.684)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/byte/one.Binary (0.626)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/bool/false.Compact (0.617)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/double/NaN.Compact (0.654)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/binary/spaces.Binary (0.727)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/i16/one.Compact (0.730)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/i32/zero.SimpleJson (0.412)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/string/upper.Compact (0.389)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/i32/neg_one.Compact (0.420)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/NaN.Compact (0.410)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/i16/max.Binary (0.421)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/byte/min.Compact (0.391)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/neg_zero.Compact (0.384)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/bool/true.SimpleJson (0.396)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/string/utf8.Compact (0.396)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/neg_zero.SimpleJson (0.381)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/float/zero.Binary (0.426)
      ✓ Pass: thrift/conformance:round-trip-test - cpp2/RoundTripTest/i32/zero.Binary (0.411)
  Summary
    Pass: 204
  ListingSuccess: 1

Reviewed By: iahs

Differential Revision: D24749395

fbshipit-source-id: 7ff02d7545269aa3bdbeda4065a403c6d96aca66